Mode call-graph construction
